HobbyFragment added and Adjustments
